// client/src/app/view/SF/SF00101/model/SF001_DataTable.ts
/**
 * Created by ASUS on 6/5/2017.
 */
export class DataTableModel {
    invoiceDate: string;

    dealCode: string;

    dealName: string;

    productType: number;

    productTypeDisplay: string;

    numberOfOrder: number;

    unitPrice: string;

    amountOfMoney: number;

    shapeId: number;
}